Added external files
I added all the images and the .css file(s) that will be needed for my
web theme to be displayed correctly. I ONLY added the files to the
project, I did NOT make any changes to the code. I want to commit this
and publish it to see if it still works.