Added Folder and new Class(SimpleHash)
I have added the simpleHash class to the project ot generate the hashes
using the salt and hash method. No column for the salt will be necessary
as this should be able to append the salt to the hash and because the
hash is always the same length it can get the salt from the end of the
hash and then verrify the password that was entered. I only need to
implement a stored procedure to get the hash if the user exists.